# swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Funcao": {
				"enum": [
					"COMPRADOR",
					"VISUALIZADOR",
					"ADMINISTRADOR"
				],
				"type": "string"
			},
			"Pick_IUser.Exclude_keyofIUser.id__": {
				"properties": {
					"ativo": {
						"type": "boolean"
					},
					"emailAcesso": {
						"type": "string"
					},
					"funcao": {
						"$ref": "#/components/schemas/Funcao"
					},
					"linkFoto": {
						"type": "string"
					},
					"nome": {
						"type": "string"
					},
					"responsavel": {
						"properties": {
							"nome": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"nome",
							"id"
						],
						"type": "object"
					},
					"sobrenome": {
						"type": "string"
					},
					"gruposVinculados": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"LojasVinculadas": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"ativo",
					"emailAcesso",
					"funcao",
					"linkFoto",
					"nome",
					"responsavel",
					"sobrenome",
					"gruposVinculados",
					"LojasVinculadas"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IUser.id_": {
				"$ref": "#/components/schemas/Pick_IUser.Exclude_keyofIUser.id__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"UpdateUserRequest": {
				"$ref": "#/components/schemas/Omit_IUser.id_"
			},
			"Response__user_63_-UpdateUserRequest__": {
				"properties": {
					"error": {},
					"data": {
						"anyOf": [
							{
								"properties": {
									"user": {
										"$ref": "#/components/schemas/UpdateUserRequest"
									}
								},
								"type": "object"
							},
							{}
						]
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"code"
				],
				"type": "object"
			},
			"Pick_IUser.Exclude_keyofIUser.gruposVinculados-or-LojasVinculadas-or-ativo-or-responsavel__": {
				"properties": {
					"id": {
						"type": "string"
					},
					"emailAcesso": {
						"type": "string"
					},
					"funcao": {
						"$ref": "#/components/schemas/Funcao"
					},
					"linkFoto": {
						"type": "string"
					},
					"nome": {
						"type": "string"
					},
					"sobrenome": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"emailAcesso",
					"funcao",
					"linkFoto",
					"nome",
					"sobrenome"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IUser.gruposVinculados-or-LojasVinculadas-or-ativo-or-responsavel_": {
				"$ref": "#/components/schemas/Pick_IUser.Exclude_keyofIUser.gruposVinculados-or-LojasVinculadas-or-ativo-or-responsavel__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"Item": {
				"$ref": "#/components/schemas/Omit_IUser.gruposVinculados-or-LojasVinculadas-or-ativo-or-responsavel_"
			},
			"GetUsersResponse": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/Item"
						},
						"type": "array"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"items",
					"totalPages"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Response_GetUsersResponse_": {
				"properties": {
					"error": {},
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/GetUsersResponse"
							},
							{}
						]
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"code"
				],
				"type": "object"
			},
			"IUser": {
				"properties": {
					"ativo": {
						"type": "boolean"
					},
					"emailAcesso": {
						"type": "string"
					},
					"funcao": {
						"$ref": "#/components/schemas/Funcao"
					},
					"id": {
						"type": "string"
					},
					"linkFoto": {
						"type": "string"
					},
					"nome": {
						"type": "string"
					},
					"responsavel": {
						"properties": {
							"nome": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"nome",
							"id"
						],
						"type": "object"
					},
					"sobrenome": {
						"type": "string"
					},
					"gruposVinculados": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"LojasVinculadas": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"ativo",
					"emailAcesso",
					"funcao",
					"id",
					"linkFoto",
					"nome",
					"responsavel",
					"sobrenome",
					"gruposVinculados",
					"LojasVinculadas"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Response__user_63_-IUser__": {
				"properties": {
					"error": {},
					"data": {
						"anyOf": [
							{
								"properties": {
									"user": {
										"$ref": "#/components/schemas/IUser"
									}
								},
								"type": "object"
							},
							{}
						]
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"code"
				],
				"type": "object"
			},
			"Response": {
				"properties": {
					"error": {},
					"data": {
						"nullable": true
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"code"
				],
				"type": "object"
			},
			"Response__idUsuario-string__": {
				"properties": {
					"error": {},
					"data": {
						"anyOf": [
							{
								"properties": {
									"idUsuario": {
										"type": "string"
									}
								},
								"required": [
									"idUsuario"
								],
								"type": "object"
							},
							{}
						]
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"code"
				],
				"type": "object"
			},
			"Pick_IUser.Exclude_keyofIUser.ativo-or-id-or-linkFoto__": {
				"properties": {
					"emailAcesso": {
						"type": "string"
					},
					"funcao": {
						"$ref": "#/components/schemas/Funcao"
					},
					"nome": {
						"type": "string"
					},
					"responsavel": {
						"properties": {
							"nome": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"nome",
							"id"
						],
						"type": "object"
					},
					"sobrenome": {
						"type": "string"
					},
					"gruposVinculados": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"LojasVinculadas": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"emailAcesso",
					"funcao",
					"nome",
					"responsavel",
					"sobrenome",
					"gruposVinculados",
					"LojasVinculadas"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IUser.ativo-or-id-or-linkFoto_": {
				"$ref": "#/components/schemas/Pick_IUser.Exclude_keyofIUser.ativo-or-id-or-linkFoto__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"ICreateUserRequest": {
				"$ref": "#/components/schemas/Omit_IUser.ativo-or-id-or-linkFoto_"
			},
			"Response__id-string__": {
				"properties": {
					"error": {},
					"data": {
						"anyOf": [
							{
								"properties": {
									"id": {
										"type": "string"
									}
								},
								"required": [
									"id"
								],
								"type": "object"
							},
							{}
						]
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"code"
				],
				"type": "object"
			},
			"IStores": {
				"properties": {
					"nomeLoja": {
						"type": "string"
					},
					"codLoja": {
						"type": "string"
					},
					"responsavel": {
						"properties": {
							"id": {
								"type": "string"
							},
							"nome": {
								"type": "string"
							}
						},
						"required": [
							"id",
							"nome"
						],
						"type": "object"
					},
					"gruposVinculados": {
						"items": {
							"properties": {
								"id": {
									"type": "string"
								},
								"codGrupo": {
									"type": "string"
								},
								"nomeGrupo": {
									"type": "string"
								}
							},
							"required": [
								"id",
								"codGrupo",
								"nomeGrupo"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"nomeLoja",
					"codLoja",
					"responsavel"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Response__store-IStores__": {
				"properties": {
					"error": {},
					"data": {
						"anyOf": [
							{
								"properties": {
									"store": {
										"$ref": "#/components/schemas/IStores"
									}
								},
								"required": [
									"store"
								],
								"type": "object"
							},
							{}
						]
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"code"
				],
				"type": "object"
			},
			"IGroups": {
				"properties": {
					"nomeGrupo": {
						"type": "string"
					},
					"codGrupo": {
						"type": "string"
					},
					"responsavel": {
						"properties": {
							"id": {
								"type": "string"
							},
							"nome": {
								"type": "string"
							}
						},
						"required": [
							"id",
							"nome"
						],
						"type": "object"
					},
					"lojasVinculadas": {
						"items": {
							"properties": {
								"id": {
									"type": "string"
								},
								"codLoja": {
									"type": "string"
								},
								"nomeLoja": {
									"type": "string"
								}
							},
							"required": [
								"id",
								"codLoja",
								"nomeLoja"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"nomeGrupo",
					"codGrupo",
					"responsavel"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Response__group_63_-IGroups__": {
				"properties": {
					"error": {},
					"data": {
						"anyOf": [
							{
								"properties": {
									"group": {
										"$ref": "#/components/schemas/IGroups"
									}
								},
								"type": "object"
							},
							{}
						]
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"code"
				],
				"type": "object"
			},
			"Response__groups-Array_IGroups___": {
				"properties": {
					"error": {},
					"data": {
						"anyOf": [
							{
								"properties": {
									"groups": {
										"items": {
											"$ref": "#/components/schemas/IGroups"
										},
										"type": "array"
									}
								},
								"required": [
									"groups"
								],
								"type": "object"
							},
							{}
						]
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"code"
				],
				"type": "object"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "template-api-node",
		"version": "1.0.0",
		"description": "template api node + typescript",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Caio Alves"
		}
	},
	"paths": {
		"/users/patch/{id}": {
			"patch": {
				"operationId": "Patch",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response__user_63_-UpdateUserRequest__"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserRequest"
							}
						}
					}
				}
			}
		},
		"/users/get": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response_GetUsersResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "nome",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "emailAcesso",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "funcao",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/Funcao"
						}
					},
					{
						"in": "query",
						"name": "ativo",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "direction",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					}
				]
			}
		},
		"/users/get/{id}": {
			"get": {
				"operationId": "GetUserPerID",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response__user_63_-IUser__"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/delete/{id}": {
			"delete": {
				"operationId": "DeleteUserLogic",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/create": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response__idUsuario-string__"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateUserRequest"
							}
						}
					}
				}
			}
		},
		"/stores/create": {
			"post": {
				"operationId": "PostStores",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response__id-string__"
								}
							}
						}
					}
				},
				"tags": [
					"Stores"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IStores"
							}
						}
					}
				}
			}
		},
		"/stores/patch/{id}": {
			"patch": {
				"operationId": "PatchStores",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response__store-IStores__"
								}
							}
						}
					}
				},
				"tags": [
					"Stores"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IStores"
							}
						}
					}
				}
			}
		},
		"/groups/create": {
			"post": {
				"operationId": "PostGroup",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response__id-string__"
								}
							}
						}
					}
				},
				"tags": [
					"Groups"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IGroups"
							}
						}
					}
				}
			}
		},
		"/groups/patch/{id}": {
			"patch": {
				"operationId": "PatchGroups",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response__group_63_-IGroups__"
								}
							}
						}
					}
				},
				"tags": [
					"Groups"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IGroups"
							}
						}
					}
				}
			}
		},
		"/groups/get": {
			"get": {
				"operationId": "GetGroups",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response__groups-Array_IGroups___"
								}
							}
						}
					}
				},
				"tags": [
					"Groups"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "nomeGrupo",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "codGrupo",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "nomeResponsavel",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}